==> 01_crawl.ipynb <==
from google_play_scraper import Sort, reviews, app
import pandas as pd
import time
from tqdm import tqdm, trange

roblox_dfs = []
token = None

for _ in trange(10000):
    review, token = reviews(
        'com.roblox.client',
        lang = 'en',
        country = 'us',
        sort = Sort.MOST_RELEVANT,
        count = 100,
        continuation_token = token
    )

    roblox_dfs.append(pd.DataFrame(review))

len(roblox_dfs)

roblox_df = pd.concat(roblox_dfs).reset_index(drop = True)

roblox_df.to_csv('./roblox.csv')

zepeto_dfs = []
token = None

for _ in trange(10000):
    review, token = reviews(
        'me.zepeto.main',
        lang = 'en',
        country = 'us',
        sort = Sort.MOST_RELEVANT,
        count = 100,
        continuation_token = token
    )

    zepeto_dfs.append(pd.DataFrame(review))

zepeto_df = pd.concat(zepeto_dfs).reset_index(drop = True)

zepeto_df.to_csv('./zepeto.csv')

roblox_df

zepeto_df

roblox_df = roblox_df.reset_index(drop = True)

roblox_df.loc[0:99999].to_csv('./roblox1.csv')
roblox_df.loc[100000:199999].to_csv('./roblox2.csv')
roblox_df.loc[200000:299999].to_csv('./roblox3.csv')
roblox_df.loc[300000:399999].to_csv('./roblox4.csv')
roblox_df.loc[400000:].to_csv('./roblox5.csv')
