==> 02_1_tfidf.ipynb <==
from keybert import KeyBERT
from nltk.tag import pos_tag
from nltk.tokenize import word_tokenize
import pandas as pd
import torch
import os
from tqdm import tqdm, trange
from random import randint
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.stem import PorterStemmer
from nltk.tag import pos_tag

vectorizer = TfidfVectorizer()
lemmatizer = WordNetLemmatizer()
stopword_list = stopwords.words('english')

tqdm.pandas()

def morphs(text, noun = True, verb = False, adjective = False, adverb = False):
    poses = pos_tag(text, tagset = 'universal')
    filters = []

    if noun:
        filters.append('NOUN')
    if verb:
        filters.append('VERB')
    if adjective:
        filters.append('ADJ')
    if adverb:
        filters.append('ADV')

    return [pos[0] for pos in poses if pos[1] in filters]

df1 = pd.read_csv('./datasets/original/roblox1.csv', index_col = 0, low_memory = False)
df2 = pd.read_csv('./datasets/original/roblox2.csv', index_col = 0, low_memory = False)
df3 = pd.read_csv('./datasets/original/roblox3.csv', index_col = 0, low_memory = False)
df4 = pd.read_csv('./datasets/original/roblox4.csv', index_col = 0, low_memory = False)
df5 = pd.read_csv('./datasets/original/roblox5.csv', index_col = 0, low_memory = False)
df6 = pd.read_csv('./datasets/original/zepeto.csv', index_col = 0, low_memory = False)

df = pd.concat([df1, df2, df3, df4, df5, df6]).reset_index(drop = True)

for df in [df1, df2, df3, df4, df5, df6]:
    df['keywords'] = df['content'].progress_apply(word_tokenize)                                                             # Tokenize
    df['keywords'] = df['keywords'].progress_apply(lambda x : [y for y in x if y.isalpha()])                                  # exclude if numeric
    df['keywords'] = df['keywords'].progress_apply(lambda x : [lemmatizer.lemmatize(y) for y in x])                           # lemmatize
    df['keywords'] = df['keywords'].progress_apply(lambda x : list(set(x)))                                                   # drop duplicate
    df['keywords'] = df['keywords'].progress_apply(lambda x : [y for y in x if len(y) >= 3 and len(y) <= 15])                 # 3 <= len(keyword) <= 15
    df['keywords'] = df['keywords'].progress_apply(morphs, noun = True, verb = True, adjective = False, adverb = False)       # select noun
    df['keywords'] = df['keywords'].progress_apply(lambda x : [y for y in x if y not in stopword_list])                       # stopwords
    df['keywords'] = df['keywords'].progress_apply(lambda x : ' '.join(x))

df1.to_csv('./datasets/tfidf/roblox1.csv')
df2.to_csv('./datasets/tfidf/roblox2.csv')
df3.to_csv('./datasets/tfidf/roblox3.csv')
df4.to_csv('./datasets/tfidf/roblox4.csv')
df5.to_csv('./datasets/tfidf/roblox5.csv')
df6.to_csv('./datasets/tfidf/zepeto.csv')
